- Movendo métodos para as viewmodels. - Refatorações.
