+++ prj/app/src/main/java/com/brenobatista/desafiozup/ViewModels/BaseViewModel.kt
@@ -1,0 +1,13 @@
+package com.brenobatista.desafiozup.ViewModels
+
+import androidx.lifecycle.ViewModel
+import com.brenobatista.desafiozup.Data.ReposAdapter
+import com.brenobatista.desafiozup.Services.createService
+
+open class BaseViewModel : ViewModel() {
+
+    val reposAdapter = ReposAdapter()
+    var filterAdapter = ReposAdapter()
+
+    val service = createService()
+}
